Revert the folder move
