+++ src/main/java/redgatesqlci/Utils.java
@@ -1,4 +1,4 @@
-package redgatedlmautomation;
+package redgatesqlci;
 
 import hudson.EnvVars;
 import hudson.Launcher;
