+++ src/main/resources/redgatesqlci/BuildBuilder/global.jelly
