+++ src/main/resources/redgatesqlci/PublishBuilder/config.jelly
