+++ src/main/resources/redgatesqlci/PublishBuilder/global.jelly
