+++ src/main/resources/redgatesqlci/SyncBuilder/config.jelly
