+++ src/main/resources/redgatesqlci/SyncBuilder/global.jelly
