+++ src/main/resources/redgatesqlci/TestBuilder/global.jelly
